+++ core/app/app_outgoing.py
@@ -8,7 +8,7 @@ from prometheus_client import (
 )
 import ujson
 
-from .app_elasticsearch import (
+from .elasticsearch import (
     ESMetricsUnavailable,
     es_bulk,
     es_feed_activities_total,
@@ -25,23 +25,23 @@ from .app_elasticsearch import (
     refresh_index,
 )
 
-from .app_feeds import (
+from .feeds import (
     parse_feed_config,
 )
-from .app_logger import (
+from .logger import (
     get_root_logger,
     logged,
 )
-from .app_metrics import (
+from .metrics import (
     metric_counter,
     metric_inprogress,
     metric_timer,
     get_metrics,
 )
-from .app_raven import (
+from .raven import (
     get_raven_client,
 )
-from .app_redis import (
+from .redis import (
     redis_get_client,
     acquire_and_keep_lock,
     set_feed_updates_seed_url_init,
@@ -51,7 +51,7 @@ from .app_redis import (
     redis_set_metrics,
     set_feed_status,
 )
-from .app_utils import (
+from .utils import (
     Context,
     async_repeat_until_cancelled,
     cancel_non_current_tasks,
